[test] move operations tests
