+++ tests/integration/nano/test_sign/operations/test_sign_ballot.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
-# Copyright 2023 Functori <contact@functori.com>
+# Copyright 2024 Functori <contact@functori.com>
+# Copyright 2024 Trilitech <contact@trili.tech>
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,28 +14,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Check signing origination"""
+"""Gathering of tests related to Ballot operations."""
 
 from pathlib import Path
 
-from utils.app import TezosAppScreen, DEFAULT_ACCOUNT
-from utils.message import Origination
+from utils.app import Screen, TezosAppScreen, DEFAULT_ACCOUNT
+from utils.message import Ballot
 
-def test_sign_origination(app: TezosAppScreen):
-    """Check signing origination"""
+def test_sign_ballot(app: TezosAppScreen):
+    """Check signing ballot"""
     test_name = Path(__file__).stem
 
-    app.setup_expert_mode()
+    app.assert_screen(Screen.HOME)
 
-    message = Origination(
+    message = Ballot(
         source = 'tz1ixvCiPJYyMjsp2nKBVaq54f6AdbV8hCKa',
-        fee = 10000,
-        counter = 2,
-        gas_limit = 3,
-        storage_limit = 4,
-        code = {'prim': 'UNPAIR'},
-        storage = {'prim': 'pair', 'args': [{'string': '1'}, {'int': 2}]},
-        balance = 500000
+        proposal = 'ProtoALphaALphaALphaALphaALphaALphaALpha61322gcLUGH',
+        ballot = 'yay',
+        period = 32
     )
 
     data = app.sign(DEFAULT_ACCOUNT,
