+++ tests/integration/nano/test_sign/operations/test_sign_proposals.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
-# Copyright 2023 Functori <contact@functori.com>
+# Copyright 2024 Functori <contact@functori.com>
+# Copyright 2024 Trilitech <contact@trili.tech>
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,26 +14,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Check signing set deposit limit"""
+"""Gathering of tests related to Proposals operations."""
 
 from pathlib import Path
 
 from utils.app import Screen, TezosAppScreen, DEFAULT_ACCOUNT
-from utils.message import SetDepositLimit
+from utils.message import Proposals
 
-def test_sign_set_deposit_limit(app: TezosAppScreen):
-    """Check signing set deposit limit"""
+def test_sign_proposals(app: TezosAppScreen):
+    """Check signing proposals"""
     test_name = Path(__file__).stem
 
     app.assert_screen(Screen.HOME)
 
-    message = SetDepositLimit(
-        source = 'tz3XeTwXXJeWNgVR3LqMcyBDdnxjbZ7TeEGH',
-        fee = 60000,
-        counter = 4,
-        gas_limit = 376,
-        storage_limit = 4,
-        limit = 20000
+    message = Proposals(
+        source = 'tz1ixvCiPJYyMjsp2nKBVaq54f6AdbV8hCKa',
+        proposals = [
+            'ProtoALphaALphaALphaALphaALphaALphaALpha61322gcLUGH',
+            'ProtoALphaALphaALphaALphaALphaALphaALphabc2a7ebx6WB'
+        ],
+        period = 32
     )
 
     data = app.sign(DEFAULT_ACCOUNT,
