+++ tests/integration/nano/test_sign/operations/test_sign_sc_rollup_originate.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
-# Copyright 2023 Functori <contact@functori.com>
+# Copyright 2024 Functori <contact@functori.com>
+# Copyright 2024 Trilitech <contact@trili.tech>
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Check signing smart rollup originate"""
+"""Gathering of tests related to Smart-rollup Originate operations."""
 
 from pathlib import Path
 from typing import List, Optional
