--- tests/integration/nano/test_sign_increase_paid_storage.py
@@ -1,50 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2023 Functori <contact@functori.com>
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-# http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Check signing increase paid storage"""
-
-from pathlib import Path
-
-from utils.app import Screen, TezosAppScreen, DEFAULT_ACCOUNT
-from utils.message import IncreasePaidStorage
-
-def test_sign_increase_paid_storage(app: TezosAppScreen):
-    """Check signing increase paid storage"""
-    test_name = Path(__file__).stem
-
-    app.assert_screen(Screen.HOME)
-
-    message = IncreasePaidStorage(
-        source = 'tz1ixvCiPJYyMjsp2nKBVaq54f6AdbV8hCKa',
-        fee = 10000,
-        counter = 2,
-        gas_limit = 3,
-        storage_limit = 4,
-        amount = 5,
-        destination = "KT18amZmM5W7qDWVt2pH6uj7sCEd3kbzLrHT"
-    )
-
-    data = app.sign(DEFAULT_ACCOUNT,
-                    message,
-                    with_hash=True,
-                    path=test_name)
-
-    app.checker.check_signature(
-        account=DEFAULT_ACCOUNT,
-        message=message,
-        with_hash=True,
-        data=data)
-
-    app.quit()
